% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{batchCorrection}
\alias{batchCorrection}
\title{Either corrects or accounts for (known) batch effects with one of several algorithms.}
\usage{
batchCorrection(
  input.obj,
  method = c("lm", "lmm", "sva", "ruv2", "ruv4", "ruv3", "bmc", "bat", "rbe", "fab",
    "pn", "svd"),
  update = TRUE
)
}
\arguments{
\item{input.obj, }{phyloseq object or numeric matrix (correct orientation is handeled internally)}

\item{method, }{algorithm to use}

\item{meta.obj, }{data.frame of covariates that is required for matrix-like inputs}
}
\value{
will return sth. at some point
}
\description{
Either corrects or accounts for (known) batch effects with one of several algorithms.
}
