% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{colinScore}
\alias{colinScore}
\title{Takes a fitted model and computes maximum correlation between covariates as return value.
Return value contains actual correlation-matrix as 'vcor' attribute.
Works for lm and lmm.}
\usage{
colinScore(model.fit)
}
\arguments{
\item{model.fit}{lm() or lmm() output}
}
\description{
Takes a fitted model and computes maximum correlation between covariates as return value.
Return value contains actual correlation-matrix as 'vcor' attribute.
Works for lm and lmm.
}
