% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datadummy}
\alias{datadummy}
\title{Mock-up microbiome abundance table and meta-data.}
\format{
A list with two elements:
\describe{
  \item{cnts}{Compositional Abundance Data}
  \item{meta}{Covariate Information}
  ...
}
}
\usage{
datadummy
}
\description{
A dataset containing pre-processed abundance table of
microbial communities and a matrix of covariate information.
}
\keyword{datasets}
