% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy.mbec}
\alias{dummy.mbec}
\title{Mock-up microbiome abundance table and meta-data.}
\format{
An mbecData object including tss and clr transformed counts:
\describe{
  \item{otu}{Compositional Abundance Data}
  \item{tss}{Compositional Abundance Data Sum-Scaled}
  \item{clr}{Compositional Abundance Data Log-Ratio Transformed}
  \item{meta}{Covariate Information}
  ...
}
}
\source{
\url{http://www.maulraus.de/}
}
\usage{
dummy.mbec
}
\description{
A dataset containing pre-processed abundance table of
microbial communities and a matrix of covariate information.
}
\keyword{datasets}
