% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy.ps}
\alias{dummy.ps}
\title{Mock-up microbiome abundance table and meta-data.}
\format{
A phyloseq object containing counts and meta-data:
\describe{
  \item{otu_table}{Compositional Abundance Data}
  \item{sam_data}{Covariate Information}
}
}
\usage{
dummy.ps
}
\description{
An artificial data-set containing pre-processed abundance table of
microbial communities and a matrix of covariate information. The data was
created using the \code{\link{mbecDummy}} function for the sole purpose of
running examples and showing the package workflow. This particular object
was then converted using \code{\link{phyloseq}}.
}
\examples{
data(dummy.ps)
}
\keyword{datasets}
