% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{LRTransform}
\alias{LRTransform}
\title{Wrapper to help perform log-ratio transformations.
Adapted from packages 'mixOmics' and 'robCompositions' to work on matrices and Phyloseq objects alike.}
\usage{
LRTransform(
  input.obj,
  meta.obj = NULL,
  method = c("none", "CLR", "ILR"),
  offset = 0
)
}
\arguments{
\item{input.obj, }{either pyhloseq-object (OTU orientation is handled) or numeric matrix (samples x features)}

\item{method, }{one of 'CLR' or 'ILR'}

\item{offset, }{optional offset in case of sparse matrix}
}
\description{
Wrapper to help perform log-ratio transformations.
Adapted from packages 'mixOmics' and 'robCompositions' to work on matrices and Phyloseq objects alike.
}
