% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecBat}
\alias{mbecBat}
\title{Combat Batch Effects (ComBat)}
\usage{
mbecBat(input.obj, model.vars)
}
\arguments{
\item{input.obj}{phyloseq object or numeric matrix (correct orientation is
handeled internally)}

\item{model.vars}{First variable determines Batch, second (optional) variable
describes study-group. ToDo: more vars and own models}
}
\value{
A matrix of batch-effect corrected counts
}
\description{
This method uses an non-/parametric empirical  Bayes framework to correct
for BEs. Described by Johnson et al. 2007 this method was initially conceived
to work with gene expression data and is part of the sva-package in R.
}
\details{
The function returns a matrix of corrected counts Correct orientation of
counts will be handled internally.
}
\keyword{BECA}
\keyword{Batch}
\keyword{Centering}
\keyword{Mean}
