% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecBox}
\alias{mbecBox}
\title{Produces box-plot + density for OTUs in phyloseq object.}
\usage{
mbecBox(
  input.obj,
  method = c("ALL", "TOP"),
  n = 10,
  model.var = "batch",
  return.data = FALSE
)
}
\arguments{
\item{input.obj, }{list(cnts, meta), phyloseq, MbecData object (correct orientation is handeled internally)}

\item{method, }{one of 'ALL' or 'TOP' for 'n' most variable features, DEFAULT is 'ALL'}

\item{n, }{number of OTUs to display for 'TOP' method}

\item{model.var, }{covariate to group by, default is batch}

\item{return.data, }{logical if TRUE returns the data.frame required for plotting (NO plotting or saving here bucko)}
}
\description{
Produces box-plot + density for OTUs in phyloseq object.
}
