% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecCorrection}
\alias{mbecCorrection}
\title{Batch Effect Correction}
\usage{
mbecCorrection(
  input.obj,
  model.vars = c("group", "batch"),
  method = c("lm", "lmm", "sva", "ruv2", "ruv4", "ruv3", "bmc", "bat", "rbe", "fab",
    "pn", "svd"),
  update = TRUE
)
}
\arguments{
\item{input.obj, }{mbecData object or numeric matrix (correct orientation is handled internally)}

\item{model.vars}{two covariates of interest to select by first variable selects panels and second one determines coloring}

\item{method, }{algorithm to use}
}
\value{
an object of class MbecDataSet
}
\description{
Either corrects or accounts for (known) batch effects with one of several algorithms.
The methods 'lm, lmm, sva, ruv-2 and ruv-4" assess batch effects and return STH. I GUESS...
The remaining methods return the batch-corrected counts.
}
\details{
Assessment Methods

Linear (Mixed) Models:

Surrogat3e variable Analysis (SVA):

Remove unwanted Variation 2 (RUV-2):

Remove Unwanted Variation 4 (RUV-4):

Correction Methods
Remove Unwanted Variation 3 (RUV-3):

Batch Mean Centering (BMC):

Combat Batch Effects (ComBat):

Remove Batch Effects (RBE):

FAbatch: maybe just remove this - or keep in case of non-microbiome data

Percentile Normalization (PN):

Support Vector Decomposition (SVD):

The function returns either a plot-frame or the finished ggplot object. Input for th data-set can
be an MbecData-object, a phyloseq-object or a list that contains counts and covariate data. The
covariate table requires an 'sID' column that contains sample IDs equal to the sample naming in
the counts table. Correct orientation of counts will be handled internally.
}
\examples{
This will return p-value for the linear model fit
\dontrun{val.score <- mbecLM(input.obj, model.vars=c("group","batch"), method="lm")}
}
\keyword{Assessment}
\keyword{Batch-Effect}
\keyword{Correction}
