% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{MbecData}
\alias{MbecData}
\title{Define MbecData-class}
\usage{
MbecData(
  cnt_table = NULL,
  meta_data = NULL,
  type = character(),
  log = character(),
  tax_table = NULL,
  phy_tree = NULL,
  refseq = NULL,
  assessments = list(),
  corrections = list(),
  tss = NULL,
  clr = NULL
)

MbecData(
  cnt_table = NULL,
  meta_data = NULL,
  type = character(),
  log = character(),
  tax_table = NULL,
  phy_tree = NULL,
  refseq = NULL,
  assessments = list(),
  corrections = list(),
  tss = NULL,
  clr = NULL
)
}
\arguments{
\item{cnt_table}{either class phyloseq or a matrix of counts}

\item{meta_data}{A table with covariate information, whose row-names correspond to sample-IDs}

\item{type}{string type that describes the data, e.g., raw, processed, ..}

\item{log}{log that will be filled by the other package functions}

\item{tax_table}{taxonomic table from phyloseq as optional input}

\item{phy_tree}{phylogenetic tree as optional input}

\item{refseq}{reference sequences as optional input}
}
\value{
produces an R-object of type MbecData
}
\description{
An extension of phyloseq-class that contains the additional fields 'type', 'log' and
'transformations' to accommodate MBECS functionality.

Constructor for the package class MbecData that takes a single input object of class phyloseq or a matrix of counts
and a data-frame of covariate variables for model-building.
}
\section{Slots}{

\describe{
\item{\code{type}}{User defined denominator for this data, e.g., 'RAW', 'normalized'.}

\item{\code{log}}{log that will be filled by the other package functions}

\item{\code{otu_table}}{Class phyloseq::otu_table, (usually sparse) matrix of
abundance values.}

\item{\code{sample_data}}{Dataframe of covariate variables.}

\item{\code{tax_table}}{taxonomic table from phyloseq as optional input}

\item{\code{phy_tree}}{phylogenetic tree as optional input}

\item{\code{refseq}}{reference sequences as optional input}

\item{\code{assessments}}{A list for the results of BEAAs.}

\item{\code{corrections}}{A list for the results of BECAs.}

\item{\code{tss}}{Total-sum-squared features matrix.}

\item{\code{clr}}{Cumulative log-ratio transformed feature matrix.}

\item{\code{assessments}}{A list for the results of BEAAs.}

\item{\code{corrections}}{A list for the results of BECAs.}

\item{\code{tss}}{Total-sum-squared features matrix.}

\item{\code{clr}}{Cumulative log-ratio transformed feature matrix.}
}}

\examples{
# use constructor with default parameters to create object from count-matrix
# and meta-data table.
mbec.obj <- MbecData(cnt_table=datadummy$cnts, meta_data = datadummy$meta)
# use constructor with default parameters to create object from count-matrix
# and meta-data table.
mbec.obj <- MbecData(cnt_table=datadummy$cnts, meta_data = datadummy$meta)
}
\keyword{Class}
\keyword{Constructor}
\keyword{MBECS}
