% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecDeflate}
\alias{mbecDeflate}
\title{Calculate matrix residuals}
\usage{
mbecDeflate(input.mtx, t)
}
\arguments{
\item{input.mtx}{A matrix of counts (samples x features).}

\item{t}{An sxf matrix object of latent components.}
}
\value{
A matrix of residual counts of same size and orientation as the
input.
}
\description{
Internal function that performs matrix deflation to remove latent components
from a sxf oriented matrix to produce the residual matrix.
}
\keyword{components}
\keyword{deflation}
\keyword{latent}
\keyword{matrix}
\keyword{residual}
