% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mbecDummy}
\alias{mbecDummy}
\title{Creates a dummy data-set with abundance matrix and meta-data.}
\usage{
mbecDummy(n.otus = 500, n.samples = 40)
}
\arguments{
\item{n.otus}{Integer to determine number of features to "simulate".}

\item{n.samples}{Even integer to set number of samples to "simulate".}
}
\value{
A list object that contains the made up abundance and the
accompanying meta-data.
}
\description{
For given number of otus and samples this will create mockup microbiome data
that contains systematic and non-systematic batch effects. Comes with meta
data that denotes study groups and batches. The replicate column is fake
and only used to test RUV-implementations.
}
\details{
'Group' and 'batch' variables are actually taken into account in data
creation, but only to the degree that the random draws for values are
performed with different parameters respectively.

THIS HAS ONLY A CONCEPTUAL SIMILARITY TO MICROBIOME DATA AT BEST AND IS IN
NO WAY USEFUL OTHER THAN TESTING PACKAGE FUNCTIONS AND VISUALIZING WORKFLOWS!
}
\examples{
dummy.list <- mbecDummy(n.otus=100, n.samples=30)
}
