% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecFAB}
\alias{mbecFAB}
\title{FAbatch}
\usage{
mbecFAB(input.obj, model.vars)
}
\arguments{
\item{input.obj}{phyloseq object or numeric matrix (correct orientation is
handeled internally)}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring}
}
\value{
A matrix of batch-effect corrected counts
}
\description{
Implemented in the bapred-package this method is based on the approach
described in Hornung et al. 2017. "It is a combination of two commonly used
approaches: location-and-scale adjustment and data cleaning by adjustment
for distortions due to latent factors." HOWEVER, since it can't handle the
zero inflated counts of compositional microbiome-data --> maybe just remove
this - or keep in case of non-microbiome data?!
}
\details{
The function returns a matrix of corrected counts Correct orientation of
counts will be handled internally.
}
\keyword{BECA}
\keyword{FABatch}
\keyword{Hornung}
