% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecGetData,list-method}
\alias{mbecGetData,list-method}
\title{Mbec-Data Getter List}
\usage{
\S4method{mbecGetData}{list}(input.obj, orientation = "fxs", required.col = NULL)
}
\arguments{
\item{input.obj}{phyloseq-object}

\item{orientation, }{Select either 'fxs' or 'sxf' to retrieve features in rows or columns respectively}

\item{required.col}{Vector of column names that are required from the covariate-table.}
}
\value{
A list that contains count-matrix (in chosen orientation) and meta-data table.
}
\description{
This function extracts abundance matrix and meta-data in the chosen orientation from list input.
The implementation for lists basically performs some type checks on the input and then compares
col/row-names of abundance table and meta-data to infer orientation and return the user specified
direction
}
\details{
The parameter 'orientation' determines if the output has features as columns (sxf) or if the
columns contain samples (fxs). This is mainly used to retrieve correctly oriented matrices for
the different analysis and correction functions.

The parameter 'required.col' is a vector of column names (technically positions would work) from
meta-data, that are required for the analysis at hand. The function actually only checks if
they are present in the data, but it will return the whole meta-frame.
}
\examples{
# This will return the abundance matrix with samples as rows and check for the presence of
# variables 'group' and 'batch' in the meta-data.
list.obj <- mbecGetData(input.obj=datadummy, orientation="sxf",
    required.col=c("group","batch"))

# This will return the abundance matrix with samples as columns and check for the presence
# of variables 'group' and 'batch' in the meta-data.
list.obj <- mbecGetData(input.obj=datadummy, orientation="fxs",
    required.col=c("group","batch"))
}
\keyword{Getter}
\keyword{MBECS}
