% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecGetData}
\alias{mbecGetData}
\title{This function is a helper tor retrieve cnts (otu_table) in desired orientation,
i.e., features in rows (fxs) or in columns (sxf).}
\usage{
mbecGetData(input.obj, orientation = "fxs", required.col = NULL)
}
\arguments{
\item{input.obj, }{can be either 'phyloseq' or 'MbecData' class}

\item{orientation, }{select either 'fxs' or 'sxf' to retrieve features in rows or columns respectively}
}
\value{
list(), containing count.matrix in first and meta-data in second slot
}
\description{
This function is a helper tor retrieve cnts (otu_table) in desired orientation,
i.e., features in rows (fxs) or in columns (sxf).
}
