% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecHeat}
\alias{mbecHeat}
\title{Produces box-plot + density for OTUs in phyloseq object.}
\usage{
mbecHeat(
  input.obj,
  model.vars = c("group", "batch"),
  center = TRUE,
  scale = TRUE,
  method = "TOP",
  n = 10,
  return.data = FALSE
)
}
\arguments{
\item{input.obj, }{list(cnts, meta), phyloseq, MbecData object (correct orientation is handeled internally)}

\item{center, }{flag to activate centering}

\item{scale, }{flag to activate scaling}

\item{method, }{one of 'ALL' or 'TOP' or a vector of feature names}

\item{return.data, }{logical if TRUE returns the data.frame required for plotting (NO plotting or saving here bucko)}

\item{grouping, }{covariate to group by, default is batch}
}
\description{
Produces box-plot + density for OTUs in phyloseq object.
}
