% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecLM}
\alias{mbecLM}
\title{DEPRECATED
Helper function that fits lm/lmm with covariates 'treatment' and 'batch' to every feature in the dataset}
\usage{
mbecLM(
  input.obj,
  method = c("auto", "lm", "lmm"),
  grouping = c("batch", "group")
)
}
\arguments{
\item{input.obj, }{mbecData object or numeric matrix (correct orientation is handeled internally)}

\item{method, }{either 'lm' or 'lmm' for linear (mixed) models or 'auto' to detect correct method (not implemented yet)}
}
\value{
vector of fdr corrected p-values that show significance of treatment for every feature
}
\description{
DEPRECATED
Helper function that fits lm/lmm with covariates 'treatment' and 'batch' to every feature in the dataset
}
