% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecMixedVariance}
\alias{mbecMixedVariance}
\title{Extract Linear-Mixed Models Variance Components}
\usage{
mbecMixedVariance(model.fit)
}
\description{
Extract Linear-Mixed Models Variance Components
}
