% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecModelVariance}
\alias{mbecModelVariance}
\title{Wrapper for calculation of variance statistics - Handles erros, validation, iteration and model selection}
\usage{
mbecModelVariance(
  input.obj,
  model.vars = character(),
  method = c("lm", "lmm", "rda", "pvca"),
  type = "NONE"
)
}
\arguments{
\item{input.obj, }{list(cnts, meta), phyloseq, MbecData object (correct orientation is handeled internally)}

\item{model.vars, }{vector of covariates to include in model-construction}

\item{method, }{WHEN 'm.formula' works, this will become obsolete}

\item{type, }{creates a column with that string in the output df - to keep track of cnt-source}

\item{m.formula, }{EVENTUALLY this will be used to set-up the model and chack covariate availability}
}
\value{
df that contains proportions of variance for given covariates in every feature
}
\description{
Wrapper for calculation of variance statistics - Handles erros, validation, iteration and model selection
}
