% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_plots.R
\name{mbecMosaicPlot}
\alias{mbecMosaicPlot}
\title{Mosaic plotting function}
\usage{
mbecMosaicPlot(study.summary, model.vars)
}
\arguments{
\item{study.summary}{'mbecMosaic' output object.}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring.}
}
\value{
ggplot2 object
}
\description{
Takes data.frame from mbecMosaic and produces a ggplot2 object.
}
\examples{
# This will return a paneled plot that shows results for the variance
# assessment.
data(dummy.mbec)
mosaic.df <- mbecMosaic(input.obj=dummy.mbec, model.vars=c('group','batch'),
return.data=TRUE)
plot.mosaic <- mbecMosaicPlot(study.summary=mosaic.df,
model.vars=c('group','batch'))
}
\keyword{RLE}
\keyword{expression}
\keyword{log}
\keyword{relative}
