% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecPCTest}
\alias{mbecPCTest}
\title{Compare Distributions of Principal Component Values per Group}
\usage{
mbecPCTest(plot.df, pca.axes, model.vars, return.table = TRUE)
}
\arguments{
\item{plot.df}{The plotting data.frame from mbecPCA function}

\item{pca.axes}{The vector the denotes the selected axes}

\item{model.vars}{The selected covariates. ToDo: fix PCA to work well with single variable}

\item{return.table}{DEFAULT is TRUE, if set to FALSE the function will return
a list of length 2 that contains lists with result objects of 'ks.test()'
named like the performed comparisons i.e. Batch1 to Batch2, etc..}
}
\value{
A data.frame with #principal components rows and n!/(2!(n-2)!)
.columns
}
\description{
For the selected Principal Components in PCA-plot perform Kolmogorov-Smirnov
test to evaluate if distributions of the batches (or any variable really) are
statistically significantly different.
}
\keyword{uppercase}
