% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecPlots}
\alias{mbecPlots}
\title{DEPRECATED
Helper function to put together, display and save plots that were create with MBEC.}
\usage{
mbecPlots(
  data.obj,
  p.type = NULL,
  grouping = "batch",
  f.title = NULL,
  save.plot = FALSE
)
}
\arguments{
\item{data.obj, }{any output of MBEC's plotting functions.}

\item{p.type, }{which kind of plot to expect}

\item{f.title, }{string filename activates plotting if provided}
}
\description{
DEPRECATED
Helper function to put together, display and save plots that were create with MBEC.
}
