% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecProcessInput,phyloseq-method}
\alias{mbecProcessInput,phyloseq-method}
\title{Mbec-Data Constructor Wrapper for Phyloseq}
\usage{
\S4method{mbecProcessInput}{phyloseq}(input.obj, required.col = NULL)
}
\arguments{
\item{input.obj}{phyloseq-object}

\item{required.col}{Vector of column names that are required from the covariate-table.}
}
\value{
An object of type MbecData that has been validated.
}
\description{
This function is a wrapper for the constructor of MbecData-objects. Given the parameter
'required.col', the function will check if the required columns are present in the data and then
return it as an MbecData-object.
}
\details{
The parameter 'required.col' is a vector of column names (technically positions would work) from
meta-data, that are required for the analysis at hand. The function actually only checks if
they are present in the data, but it will return the whole meta-frame.
}
\examples{
# This will check for the presence of variables 'group' and 'batch' in the meta-data and return
# an object of class 'MbecData'.
MbecData.obj <- mbecProcessInput(input.obj=datadummy,
    required.col=c("group","batch"))
}
\keyword{Constructor}
\keyword{MBECS}
\keyword{Phyloseq}
\keyword{Wrapper}
