% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecProcessInput}
\alias{mbecProcessInput}
\title{This function checks if the required covariate columns exist in the sample-data}
\usage{
mbecProcessInput(input.obj, required.col = NULL)
}
\arguments{
\item{input.obj, }{can be either 'phyloseq' or 'MbecData' class}

\item{orientation, }{select either 'fxs' or 'sxf' to retrieve features in rows or columns respectively}
}
\value{
list(), containing count.matrix in first and meta-data in second slot
}
\description{
This function checks if the required covariate columns exist in the sample-data
}
