% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_reports.R
\name{mbecPVCAStatsPlot}
\alias{mbecPVCAStatsPlot}
\title{Plots variance statistics produced by 'mbecModelVariance' with method 'pvca'}
\usage{
mbecPVCAStatsPlot(pvca.obj)
}
\description{
Plots variance statistics produced by 'mbecModelVariance' with method 'pvca'
}
