% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_reports.R
\name{mbecRDAStatsPlot}
\alias{mbecRDAStatsPlot}
\title{Plots variance statistics produced by 'mbecModelVariance' with method 'rda'}
\usage{
mbecRDAStatsPlot(rda.obj)
}
\description{
Plots variance statistics produced by 'mbecModelVariance' with method 'rda'
}
