% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_reports.R
\name{mbecReport}
\alias{mbecReport}
\title{Wrapper function for the analyses required to construct preliminary or post-correction reports.}
\usage{
mbecReport(input.obj, model.vars = c("group", "batch"), return.plot = FALSE)
}
\arguments{
\item{input.obj, }{list of phyloseq objects to compare, first element is considered uncorrected data}

\item{model.vars, }{required covariates to build models}

\item{return.plot, }{TRUE will return a list of all produced plots, FALSE will start rendering the report}
}
\description{
Wrapper function for the analyses required to construct preliminary or post-correction reports.
}
