% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_reports.R
\name{mbecReportPost}
\alias{mbecReportPost}
\title{Constructs an initial report of a single data-set without comparative analyses.}
\usage{
mbecReportPost(
  input.obj,
  model.vars = c("batch", "group"),
  type = "clr",
  return.data = FALSE
)
}
\arguments{
\item{input.obj, }{list of phyloseq objects to compare, first element is considered uncorrected data}

\item{model.vars, }{required covariates to build models}

\item{type}{One of 'otu', 'tss' or 'clr' to determine the abundance matrix to use for evaluation.}

\item{return.data, }{TRUE will return a list of all produced plots, FALSE will start rendering the report}
}
\value{
either a ggplot2 object or a formatted data-frame to plot from
}
\description{
Constructs an initial report of a single data-set without comparative analyses.
}
