% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecRLE}
\alias{mbecRLE}
\title{Computes RLE for groups and batches and returns plot or data frame.}
\usage{
mbecRLE(input.obj, model.vars = c("group", "batch"), return.data = FALSE)
}
\arguments{
\item{input.obj, }{list(cnts, meta), phyloseq, MbecData object (correct orientation is handeled internally)}

\item{model.vars, }{two covariates of interest to select by first variable selects panels and second one determines coloring}

\item{return.data, }{logical if TRUE returns the data.frame required for plotting (NO plotting here bucko)}
}
\description{
Computes RLE for groups and batches and returns plot or data frame.
}
