% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecSavePlot}
\alias{mbecSavePlot}
\title{DEPRECATED
Helper function to save plots as pdf (for now maybe..) in given path with given title.}
\usage{
mbecSavePlot(plot.obj, f.title = "mbecTest")
}
\arguments{
\item{plot.obj, }{output of 'mbecPlots()'}

\item{f.title, }{filename for output, defaults to 'mbecTest'}

\item{f.path, }{save directory, defaults to workdir}
}
\description{
DEPRECATED
Helper function to save plots as pdf (for now maybe..) in given path with given title.
}
