% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_reports.R
\name{mbecSCOEFStatsPlot}
\alias{mbecSCOEFStatsPlot}
\title{Plots variance statistics produced by 'mbecModelVariance' with method silhouette-coefficient}
\usage{
mbecSCOEFStatsPlot(scoef.obj)
}
\description{
Plots variance statistics produced by 'mbecModelVariance' with method silhouette-coefficient
}
