% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecSetData}
\alias{mbecSetData}
\title{Mbec-Data Setter}
\usage{
mbecSetData(
  input.obj,
  new.cnts = NULL,
  log = character(),
  type = character(),
  update = TRUE
)
}
\arguments{
\item{input.obj}{MbecData object to change}

\item{new.cnts}{matrix-like object with same dimension as 'otu_table' in input.obj}

\item{log}{character to add to the log string}

\item{type}{character to replace type-attribute or as name tag for added count-matrix}

\item{update}{logical (TRUE) replace input fields (FALSE) add to the input.obj}
}
\value{
Input object with updated attributes.
}
\description{
This function either updates counts, type and log attributes (DEFAULT) or adds a matrix of
transformed counts to the input (update=FALSE).
}
\examples{
# This will replace the current abundance-matrix, append 'rbe' (for remove-batch-effect) to the
# log and change the type to 'rbe' to indicate current status.
MBEC.obj <- mbecSetData(input.obj=datadummy, new.cnts=datadummy$cnts,
    log='rbe',type='corrected', update=TRUE)

# This will add the corrected data to the list of transformations named like the type argument.
# The 'log' attribute will be updated, but 'type' stays the same (because the main abundnce
# matrix wasn't replaced.
MBEC.obj <- mbecSetData(input.obj=datadummy, new.cnts=datadummy$cnts,
    log='rbe',type='corrected', update=FALSE)
}
\keyword{MBECS}
\keyword{Setter}
