% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecValidateModel}
\alias{mbecValidateModel}
\title{Helper function that ensure validity of the model by testing colinearity}
\usage{
mbecValidateModel(model.fit, colinearityThreshold = 0.999)
}
\arguments{
\item{model.fit}{lm() or lmm() output}

\item{colinearityThreshold}{cut-off for model rejection}
}
\description{
Helper function that ensure validity of the model by testing colinearity
}
