% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecVarianceStatsLM}
\alias{mbecVarianceStatsLM}
\title{Model Variable Variance Extraction from LM}
\usage{
mbecVarianceStatsLM(model.fit)
}
\arguments{
\item{model.fit}{A linear model object of class 'lm'.}
}
\value{
A named row-vector, containing proportional variance for model terms.
}
\description{
For a Linear Model, this function extracts the proportion of variance
that can be explained by terms and interactions and returns a named
row-vector.
}
\details{
Linear Model: Perform an analysis of variance (ANOVA) on the model.fit and
return the Sum of squares for each term, scaled by the total sum of squares.
}
\keyword{lm}
\keyword{proportion}
\keyword{variance}
