% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecVarianceStatsLMM}
\alias{mbecVarianceStatsLMM}
\title{Model Variable Variance Extraction from LMM}
\usage{
mbecVarianceStatsLMM(model.fit)
}
\arguments{
\item{model.fit}{A linear mixed model object of class 'lmerMod'.}
}
\value{
A named row-vector, containing proportional variance for model terms.
}
\description{
For a Linear Mixed Model, this function extracts the proportion of variance
that can be explained by terms and interactions and returns a named
row-vector.
}
\details{
Linear Mixed Model: employ helper function 'mbecMixedVariance' to extract
residuals, random effects and fixed effects components from the model. The
components are then transformed to reflect explained proportions of variance
for the model coefficients. The function implements transformation for
varying coefficients as well, but NO ADJUSTMENT for single or multiple
coefficients at this point.
}
\keyword{lmm}
\keyword{proportion}
\keyword{variance}
