% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_reports.R
\name{mbecVarianceStatsPlot}
\alias{mbecVarianceStatsPlot}
\title{Covariate-Variances as modelled by lm/lmm will be displayed as box-plots. Both preliminary and post statistics.}
\usage{
mbecVarianceStatsPlot(variance.obj)
}
\arguments{
\item{variance.obj, }{list or single output of 'mbecVarianceStats' with method lm}
}
\description{
Covariate-Variances as modelled by lm/lmm will be displayed as box-plots. Both preliminary and post statistics.
}
