% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{percentileNorm}
\alias{percentileNorm}
\title{The function applies normalization to every batch of a case/control study, is called by 'batchCorrection()'
group/treatment vector is limited to two classes, i.e., case/control study}
\usage{
percentileNorm(cnts, meta)
}
\arguments{
\item{cnts, }{matrix of cnts in sxf orientation}

\item{meta, }{df that requires columns 'group' and 'batch'}
}
\description{
The function applies normalization to every batch of a case/control study, is called by 'batchCorrection()'
group/treatment vector is limited to two classes, i.e., case/control study
}
