% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{poscore}
\alias{poscore}
\title{Implementation of Claire Duvallet's 'percentileofscore()' for python
Will return values in the interval \link{0,100} and 100 if reference vector has zero length.}
\usage{
poscore(cnt.vec, cnt, type = c("rank", "weak", "strict", "mean"))
}
\arguments{
\item{cnt.vec, }{vector of cnts that acts as reference for score calculation}

\item{cnt, }{value to calculate score for}

\item{type, }{one of 'rank', 'weak', 'strict' or 'mean' to determine how score is calculated}
}
\description{
Implementation of Claire Duvallet's 'percentileofscore()' for python
Will return values in the interval \link{0,100} and 100 if reference vector has zero length.
}
