% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{theme_mosaic}
\alias{theme_mosaic}
\title{MOSAIC theme for ggplot2 output.}
\usage{
theme_mosaic(
  base_size = 18,
  base_family = "Myriad Pro",
  legend_position = "top"
)
}
\arguments{
\item{base_size}{set base font size - default is 18}

\item{base_family}{set base font family - default is Myriad Pro}

\item{legend_position}{on of "top", "bottom", "left", "right" determines legend position}
}
\value{
A poster theme with legend at the bottom.
}
\description{
Default theme for plotting the packages BOX-plots.
}
\examples{
# This will return a matrix of normalised counts, according to the covariate information in meta
\dontrun{plot <- mbecPCA(input.obj, model.vars=c("group","batch"), return.data=FALSE) +
theme_box(base_size = 18, base_family = "Myriad Pro",legend_position = 'bottom')}
}
\keyword{MOSAIC}
\keyword{theme}
