% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_themes.R
\docType{data}
\name{theme_new}
\alias{theme_new}
\title{Generic theme for ggplot2 output.}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_new
}
\value{
A generic theme with legend at the bottom.
}
\description{
Uses 'theme_bw()' and puts legend at the bottom.
}
\keyword{theme_bw}
