% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_themes.R
\name{theme_poster}
\alias{theme_poster}
\title{Poster theme for ggplot2 output.}
\usage{
theme_poster(base_size = 18, base_family = "Myriad Pro")
}
\arguments{
\item{base_size}{set base font size - default is 18}

\item{base_family}{set base font family - default is Myriad Pro}
}
\value{
A poster theme with legend at the bottom.
}
\description{
Adjusts plot to use University of Luebeck (UzL) colors.
}
\examples{
# This will return a matrix of normalised counts, according to the covariate information in meta
\dontrun{plot <- ggplot2::boxplot(data) + theme_poster(base_size = 18,
base_family = "Myriad Pro")}
}
\keyword{poster}
\keyword{theme}
