% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy.mbec}
\alias{dummy.mbec}
\title{Mock-up microbiome abundance table and meta-data.}
\format{
An mbecData object including tss and clr transformed counts:
\describe{
  \item{otu}{Compositional Abundance Data}
  \item{tss}{Compositional Abundance Data Sum-Scaled}
  \item{clr}{Compositional Abundance Data Log-Ratio Transformed}
  \item{meta}{Covariate Information}
  ...
}
}
\usage{
dummy.mbec
}
\description{
An artificial data-set containing pre-processed abundance table of
microbial communities and a matrix of covariate information. The data was
created using the \code{\link{mbecDummy}} function for the sole purpose of
running examples and showing the package workflow. This particular object
was also processed with \code{\link{mbecTransform}} function in order to
generate "clr" and "tss" transformed abundance matrices.
}
\keyword{datasets}
