% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{LRTransform}
\alias{LRTransform}
\title{Log-Ratio Transformation}
\usage{
LRTransform(
  input.obj,
  method = c("none", "CLR", "ILR"),
  offset = 0,
  required.col = NULL
)
}
\arguments{
\item{input.obj}{either pyhloseq-object (OTU orientation is handled) or numeric matrix (samples x features)}

\item{method}{one of 'CLR' or 'ILR'}

\item{offset}{optional offset in case of sparse matrix}

\item{required.col}{(OPTIONAL) A vector of column names in the meta-data that need to be present. Sanity check for subsequent steps.}
}
\value{
MbecDataObject with transformed counts
}
\description{
Wrapper to help perform log-ratio transformations ,adapted from packages 'mixOmics' and
'robCompositions' to work on matrices and Phyloseq objects alike.
}
\details{
The function returns an MbecData object with tranformed counts and covariate information. Input for the data-set
can be an MbecData-object, a phyloseq-object or a list that contains counts and covariate data.
The covariate table requires an 'sID' column that contains sample IDs equal to the sample naming
in the counts table. Correct orientation of counts will be handled internally.
}
\examples{
# This will return the cumulative log-ratio transformed counts in an MbecData object
\donttest{mbec.LRT <- LRTransform(input.obj=list(counts, covariates),
method="CLR", offset=0)}

# This will return the inverse log-ratio transformed counts in an MbecData object
\donttest{mbec.LRT <- LRTransform(input.obj=list(counts, covariates),
method="ILR", offset=0)}
}
\keyword{Log}
\keyword{Ratio}
\keyword{Transformation}
