% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecBMC}
\alias{mbecBMC}
\title{Batch Mean Centering (BMC)}
\usage{
mbecBMC(input.obj, model.vars)
}
\arguments{
\item{input.obj}{phyloseq object or numeric matrix (correct orientation is
handeled internally)}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring}
}
\value{
A matrix of batch-effect corrected counts
}
\description{
For known BEs, this method takes the batches, i.e., subgroup of samples
within a particular batch, and centers them to their mean.
}
\details{
The function returns a matrix of corrected counts Correct orientation of
counts will be handled internally.
}
\keyword{BECA}
\keyword{Batch}
\keyword{Centering}
\keyword{Mean}
