% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_plots.R
\name{mbecBoxPlot}
\alias{mbecBoxPlot}
\title{Variability boxes plotting function}
\usage{
mbecBoxPlot(tmp, otu.idx, model.var, mbecCols)
}
\arguments{
\item{tmp}{Count of selected features.}

\item{otu.idx}{Index of selected Otus in the data.}

\item{model.var}{Which covariate to group Otus by.}

\item{mbecCols}{Color scheme to use for plot.}
}
\value{
ggplot2 object
}
\description{
Takes data.frame from mbecBox and produces a ggplot2 object.
}
\keyword{RLE}
\keyword{expression}
\keyword{log}
\keyword{relative}
