% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_plots.R
\name{mbecBoxPlot}
\alias{mbecBoxPlot}
\title{Variability boxes plotting function}
\usage{
mbecBoxPlot(tmp, otu.idx, model.var)
}
\arguments{
\item{tmp}{Count of selected features.}

\item{otu.idx}{Index of selected Otus in the data.}

\item{model.var}{Which covariate to group Otus by.}

\item{mbecCols}{Color scheme to use for plot.}
}
\value{
ggplot2 object
}
\description{
Takes data.frame from mbecBox and produces a ggplot2 object.
}
\examples{
# This will return a paneled plot that shows results for the variance
# assessment.
df.box <- mbecBox(input.obj=dummy.mbec, method='TOP', n=5,
model.var='batch', type="otu", return.data=TRUE)
plot.box <- mbecBoxPlot(df.box[[1]], df.box[[2]], 'batch')
}
\keyword{RLE}
\keyword{expression}
\keyword{log}
\keyword{relative}
