% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{MbecData}
\alias{MbecData}
\title{Define MbecData-class}
\usage{
MbecData(
  type = character(),
  log = character(),
  input.obj,
  meta.obj = NULL,
  required.col = c("sample", "group", "batch"),
  tax_table = NULL,
  phy_tree = NULL,
  refseq = NULL,
  list = list()
)

MbecData(
  type = character(),
  log = character(),
  input.obj,
  meta.obj = NULL,
  required.col = c("sample", "group", "batch"),
  tax_table = NULL,
  phy_tree = NULL,
  refseq = NULL,
  list = list()
)
}
\arguments{
\item{type}{string type that describes the data, e.g., raw, processed, ..}

\item{log}{log that will be filled by the other package functions}

\item{input.obj}{either class phyloseq or a matrix of counts}

\item{meta.obj}{dataframe of covariate variables}

\item{required.col}{vector of strings that denote required variables in the covariate information}

\item{tax_table}{taxonomic table from phyloseq as optional input}

\item{phy_tree}{phylogenetic tree as optional input}

\item{refseq}{reference sequences as optional input}
}
\value{
produces an R-object of type MbecData
}
\description{
An extension of phyloseq-class that contains the additional fields 'type', 'log' and
'transformations' to accommodate MBECS functionality.

Constructor for the package class MbecData that takes a single input object of class phyloseq or a matrix of counts
and a data-frame of covariate variables for model-building.
}
\keyword{Class}
\keyword{Constructor}
\keyword{MBECS}
