% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecGetData,MbecData-method}
\alias{mbecGetData,MbecData-method}
\title{Mbec-Data Getter MbecDdata}
\usage{
\S4method{mbecGetData}{MbecData}(
  input.obj,
  orientation = "fxs",
  required.col = NULL,
  type = c("otu", "ass", "cor", "clr", "tss"),
  label = character()
)
}
\arguments{
\item{input.obj}{MbecData-object}

\item{orientation, }{Select either 'fxs' or 'sxf' to retrieve features in rows or columns respectively}

\item{required.col}{Vector of column names that are required from the covariate-table.}

\item{transformation}{Default is 'NULL', otherwise give name or index of
matrix in transformation-list attribute for extraction.}
}
\value{
A list that contains count-matrix (in chosen orientation) and meta-data table.
}
\description{
This function extracts abundance matrix and meta-data in the chosen orientation from input of
class MbecData
}
\details{
The parameter 'orientation' determines if the output has features as columns (sxf) or if the
columns contain samples (fxs). This is mainly used to retrieve correctly oriented matrices for
the different analysis and correction functions.

The parameter 'required.col' is a vector of column names (technically positions would work) from
metadata, that are required for the analysis at hand. The function actually only checks if
they are present in the data, but it will return the whole meta-frame.
}
\examples{
# This will return the abundance matrix with samples as rows and check for the presence of
# variables 'group' and 'batch' in the meta-data.
list.obj <- mbecGetData(input.obj=datadummy, orientation="sxf",
    required.col=c("group","batch"))

# This will return the abundance matrix with samples as columns and check for the presence
# of variables 'group' and 'batch' in the meta-data.
list.obj <- mbecGetData(input.obj=datadummy, orientation="fxs",
    required.col=c("group","batch"))
}
\keyword{Getter}
\keyword{MBECS}
