% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_plots.R
\name{mbecHeatPlot}
\alias{mbecHeatPlot}
\title{Heatmap plotting function}
\usage{
mbecHeatPlot(center, scale, tmp.cnts, tmp.meta, model.vars)
}
\arguments{
\item{center}{Boolean flag to indicate whether data is centered or not.}

\item{scale}{Boolean flag to indicate whether data is scaled or not.}

\item{tmp.cnts}{Count values of selected features.}

\item{tmp.meta}{Covariate information for potting.}

\item{model.vars}{Two covariates of interest to select by first variable
selects panels and second one determines coloring.}
}
\value{
ggplot2 object
}
\description{
Takes data.frame from 'mbecHeat()' and produces a ggplot2 object.
}
\keyword{RLE}
\keyword{expression}
\keyword{log}
\keyword{relative}
