% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecLM}
\alias{mbecLM}
\title{Linear (Mixed) Model Feature to Batch Fit}
\usage{
mbecLM(input.obj, method = c("lm", "lmm"), model.vars = c("batch", "group"))
}
\arguments{
\item{input.obj, }{mbecData object or numeric matrix (correct orientation is handeled internally)}

\item{method, }{either 'lm' or 'lmm' for linear (mixed) models or 'auto' to detect correct method (not implemented yet)}

\item{model.vars}{two covariates of interest to select by first variable selects panels and second one determines coloring}
}
\value{
vector of fdr corrected p-values that show significance of treatment for every feature
}
\description{
Helper function that fits lm/lmm with covariates 'treatment' and 'batch' to every feature in the
data-set. Returns the fdr corrected significance value for the "treatment" variable. The method
'lm' will fit the linear model\code{y ~ model.vars[1] + model.vars[2]} and the linear mixed model
will consider the second term as random effect, i.e., \code{y ~ model.vars[1] + (1|model.vars[2])}.
}
\details{
The function returns either a plot-frame or the finished ggplot object. Input for th data-set can
be an MbecData-object, a phyloseq-object or a list that contains counts and covariate data. The
covariate table requires an 'sID' column that contains sample IDs equal to the sample naming in
the counts table. Correct orientation of counts will be handled internally.
}
\examples{
# This will return p-value for the linear model fit of every feature.
val.score <- mbecLM(input.obj=datadummy, model.vars=c("batch","group"),
method="lm")
}
\keyword{Batch}
\keyword{Linear}
\keyword{Mixed}
\keyword{Model}
\keyword{Significance}
