% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecModelVarianceLMM}
\alias{mbecModelVarianceLMM}
\title{Estimate Explained Variance with Linear Mixed Models}
\usage{
mbecModelVarianceLMM(model.form, model.vars, tmp.cnts, tmp.meta, type)
}
\arguments{
\item{model.form}{type formula for linear model}

\item{model.vars}{covariates to use for model building}

\item{tmp.cnts}{extracted count matrix}

\item{tmp.meta}{extracted covariate table}

\item{type}{String the denotes data source to keep track of process steps}
}
\value{
df that contains proportions of variance for given covariates
}
\description{
The function offers a selection of methods/algorithms to estimate the
proportion of variance that can be attributed to covariates of interest.
This shows, how much variation is explained by the treatment effect, which
proportion is introduced by processing in batches and the leftover variance,
i.e., residuals that are not currently explained. Covariates of interest
(CoI) are selected by the user and the function will incorporate them into
the model.
}
\details{
Linear Mixed Model (lmm): All but the first covariate are considered mixed
effects. A model is fitted to each OTU respectively and the proportion of
variance extracted for each covariate.
(OTU_x ~ covariate_1 + (1|covariate_2) + (1|...)).

The function returns a data-frame for further analysis - the report functions
(mbecReport and mbecReportPrelim) will automatically produce plots. Input for
the data-set can be an MbecData-object, a phyloseq-object or a list that
contains counts and covariate data. The covariate table requires an 'sID'
column that contains sample IDs equal to the sample naming in the counts
table. Correct orientation of counts will be handled internally.
}
\keyword{LM}
\keyword{Variance}
