% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecMosaic}
\alias{mbecMosaic}
\title{Produces mosaic plot to visualize sub-populations due to group-wise combinations, e.g., group x batch}
\usage{
mbecMosaic(input.obj, model.vars = c("group", "batch"), return.data = FALSE)
}
\arguments{
\item{input.obj, }{list(cnts, meta), phyloseq, MbecData object (correct orientation is handeled internally)}

\item{return.data, }{logical if TRUE returns the data.frame required for plotting (NO plotting or saving here bucko)}

\item{grouping, }{two covariates of interest to display sample distribution for}
}
\description{
Produces mosaic plot to visualize sub-populations due to group-wise combinations, e.g., group x batch
}
