% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecPCA}
\alias{mbecPCA}
\title{Creates nice PCA plot with axis-density graphs}
\usage{
mbecPCA(
  input.obj,
  model.vars = c("group", "batch"),
  pca.axes = c(1, 2),
  return.data = FALSE
)
}
\arguments{
\item{input.obj, }{list(cnts, meta), phyloseq, MbecData object (correct orientation is handeled internally)}

\item{pca.axes, }{numeric vector which axes to plot, first is X and second is Y}

\item{return.data, }{logical if TRUE returns the data.frame required for plotting (NO plotting or saving here bucko)}
}
\description{
Creates nice PCA plot with axis-density graphs
}
