% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecRLE}
\alias{mbecRLE}
\title{Relative Log Expression Plot}
\usage{
mbecRLE(
  input.obj,
  model.vars = c("batch", "group"),
  type = "clr",
  label = character(),
  return.data = FALSE
)
}
\arguments{
\item{input.obj}{list(cnts, meta), phyloseq, MbecData object (correct
orientation is handeled internally)}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring.}

\item{type}{Which abundance matrix to use for the calculation.}

\item{label}{Which corrected abundance matrix to use for analysis.}

\item{return.data}{logical if TRUE returns the data.frame required for
plotting (NO plotting here bucko)}
}
\value{
either a ggplot2 object or a formatted data-frame to plot from
}
\description{
Takes two covariates, i.e., group and batch, and computes the RLE-plot over
the grouping of the first covariate, colored by the second covariate.
Effectively illustrating the relative expression between samples from
different batches within the respective study groups. Other covariates can be
chosen as input and the function will check for factors and convert if
necessary. Categorical factors, e.g., group membership, sex and batch,
produce the best result.
}
\details{
The function returns either a plot-frame or the finished ggplot object. Input
for th data-set can be an MbecData-object, a phyloseq-object or a list that
contains counts and covariate data. The covariate table requires an 'sID'
column that contains sample IDs equal to the sample naming in the counts
table. Correct orientation of counts will be handled internally.
}
\examples{
# This will return the data.frame for plotting.
data.RLE <- mbecRLE(input.obj=dummy.mbec, type="clr",
model.vars=c('group','batch'), return.data=TRUE)

# This will return the ggplot2 object for display, saving and modification.
plot.RLE <- mbecRLE(input.obj=dummy.mbec, model.vars=c('group','batch'),
type="clr", return.data=FALSE)
}
\keyword{RLE}
\keyword{expression}
\keyword{log}
\keyword{relative}
