% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_plots.R
\name{mbecRLEPlot}
\alias{mbecRLEPlot}
\title{RLE plotting function}
\usage{
mbecRLEPlot(tmp.long, model.vars, mbecCols)
}
\arguments{
\item{tmp.long}{'mbecRLE' output}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring}

\item{mbecCols}{vector of hex-colors to use}
}
\value{
ggplot2 object
}
\description{
Takes data.frame from mbecRLE and produces a ggplot2 object.
}
\keyword{RLE}
\keyword{expression}
\keyword{log}
\keyword{relative}
