% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecRUV3}
\alias{mbecRUV3}
\title{Remove Unwanted Variation 3 (RUV-3)}
\usage{
mbecRUV3(input.obj, model.vars, nc.features = NULL)
}
\arguments{
\item{input.obj}{phyloseq object or numeric matrix (correct orientation is
handeled internally)}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring}

\item{nc.features}{(OPTIONAL) A vector of features names to be used as
negative controls in RUV-3. If not supplied, the algorithm will use an 'lm'
to find pseudo-negative controls}
}
\value{
A matrix of batch-effect corrected counts
}
\description{
This algorithm requires negative control-features, i.e., OTUs that are known
to be unaffected by the batch effect, as well as technical replicates. The
algorithm will check for the existence of a replicate column in the covariate
data. If the column is not present, the execution stops and a warning message
will be displayed.
}
\details{
The function returns a matrix of corrected counts Correct orientation of
counts will be handled internally.
}
\keyword{Assessment}
\keyword{Batch-Effect}
\keyword{Correction}
