% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecRUV4}
\alias{mbecRUV4}
\title{Remove Unwanted Variation 4 (RUV-4)}
\usage{
mbecRUV4(input.obj, model.vars, type = "clr", nc.features = NULL)
}
\arguments{
\item{input.obj}{phyloseq object or numeric matrix (correct orientation is
handeled internally)}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring}

\item{nc.features}{(OPTIONAL) A vector of features names to be used as
negative controls in RUV-3. If not supplied, the algorithm will use an 'lm'
to find pseudo-negative controls}
}
\value{
SOMETHING
}
\description{
The updated version of RUV-2 also incorporates the residual matrix
(w/o treatment effect) to estimate the unknown BEs. To that end it follows
the same procedure in case there are no negative control variables and
computes pseudo-controls from the data via l(m)m. As RUV-2, this algorithm
also uses the parameter 'k' for the number of latent factors. RUV-4 brings
the function 'getK()' that estimates this factor from the data itself. The
calculated values are however not always reliable. A value of k=0 fo example
can occur and should be set to 1 instead.
}
\details{
The output will be SOMETHING
}
\keyword{Assessment}
\keyword{Batch-Effect}
\keyword{Correction}
