% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecSCOEFStatsPlot}
\alias{mbecSCOEFStatsPlot}
\title{Plot Silhouette Coefficient}
\usage{
mbecSCOEFStatsPlot(scoef.obj)
}
\arguments{
\item{scoef.obj, }{list or single output of 'mbecVarianceStats' with method s.coef}
}
\description{
The goodness of clustering assessed by the silhouette coefficient.
It works with the output of 'mbecVarianceStats()' for the method 's.coef'. Format of this
output is a data.frame that contains a column for every model variable and as many rows as
there are features (OTUs, Genes, ..). Multiple frames may be used as input by putting them into
a list - IF the data.frames contain a column named 'type', this function will use 'facet_grid()'
to display side-by-side panels to enable easy comparison.
}
\examples{
This will return a paneled plot that shows results for three variance assessments.
\dontrun{p.sc <- mbecSCOEFStatsPlot(variance.obj=list(df1, df2, df3))}
}
\keyword{linear}
\keyword{mixed}
\keyword{models}
\keyword{plot}
\keyword{proportion}
\keyword{variance}
