% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_classes.R
\name{mbecSetData}
\alias{mbecSetData}
\title{This function either updates counts, type and log attributes or adds a matrix of transformed counts to the input}
\usage{
mbecSetData(
  input.obj,
  new.cnts = NULL,
  log = character(),
  type = character(),
  update = TRUE
)
}
\arguments{
\item{input.obj}{MbecData object to change}

\item{new.cnts}{matrix-like object with same dimension as 'otu_table' in input.obj}

\item{log}{character to add to the log string}

\item{type}{character to replace type-attribute or as name tag for added count-matrix}

\item{update}{logical (TRUE) replace input fields (FALSE) add to the input.obj}
}
\description{
This function either updates counts, type and log attributes or adds a matrix of transformed counts to the input
}
