% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_corrections.R
\name{mbecSVA}
\alias{mbecSVA}
\title{Surrogate variable Analysis (SVA)}
\usage{
mbecSVA(input.obj, model.vars)
}
\arguments{
\item{input.obj}{phyloseq object or numeric matrix (correct orientation is
handeled internally)}

\item{model.vars}{two covariates of interest to select by first variable
selects panels and second one determines coloring}

\item{nc.features}{(OPTIONAL) A vector of features names to be used as
negative controls in RUV-3. If not supplied, the algorithm will use an 'lm'
to find pseudo-negative controls}
}
\value{
SOMETHING
}
\description{
Two step approach that (1.) identify the number of latent factors to be
estimated by fitting a full-model with effect of interest and a null-model
with no effects. The function num.sv then calculates the number of latent
factors. In the next (2.) step, the sva function will estimate the surrogate
variables. And adjust for them in full/null-model . Subsequent F-test gives
significance values for each feature - these P-values and Q-values are
accounting for surrogate variables (estimated BEs).
}
\details{
The output will be SOMETHING
}
\keyword{Assessment}
\keyword{Batch-Effect}
\keyword{Correction}
