% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_helper.R
\name{mbecUpperCase}
\alias{mbecUpperCase}
\title{Capitalize Word Beginning}
\usage{
mbecUpperCase(input = character())
}
\arguments{
\item{input}{A string that needs to be Capitalized}
}
\value{
Input with first letter capitalized
}
\description{
Change the first letter of the input to uppercase. Used in plotting functions
to make covariates, i.e., axis-lables look nicer.
}
\keyword{uppercase}
