% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecValidateModel}
\alias{mbecValidateModel}
\title{Validate Linear (Mixed) Models}
\usage{
mbecValidateModel(model.fit, colinearityThreshold = 0.999)
}
\arguments{
\item{model.fit}{lm() or lmm() output}

\item{colinearityThreshold}{cut-off for model rejection}
}
\description{
A helper function that calculates the collinearity between model variables and stops execution
if the maximum value is bigger than the allowed threshold.
}
\details{
ToDo: maybe some additional validation steps and more informative output.
}
\examples{
This will return the data.frame for plotting.
\dontrun{mbecValidateModel(model.fit=MyMixedModel-obj, colinearityThreshold=0.999)}
}
\keyword{collinearity}
\keyword{model}
\keyword{validation}
