% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecVarianceStats}
\alias{mbecVarianceStats}
\title{Helper function that takes a fitted model and extracts the
fraction of variances attributable to every model-variable}
\usage{
mbecVarianceStats(model.fit)
}
\arguments{
\item{model.fit}{lm() or lmm() output}
}
\description{
Helper function that takes a fitted model and extracts the
fraction of variances attributable to every model-variable
}
