% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_analyses.R
\name{mbecVarianceStatsPlot}
\alias{mbecVarianceStatsPlot}
\title{Plot Proportion of Variance for L(M)M}
\usage{
mbecVarianceStatsPlot(variance.obj)
}
\arguments{
\item{variance.obj, }{list or single output of 'mbecVarianceStats' with method lm}
}
\description{
Covariate-Variances as modeled by linear (mixed) models will be displayed as box-plots.
It works with the output of 'mbecVarianceStats()' for methods 'lm' and 'lmm'. Format of this
output is a data.frame that contains a column for every model variable and as many rows as
there are features (OTUs, Genes, ..). Multiple frames may be used as input by putting them into
a list - IF the data.frames contain a column named 'type', this function will use 'facet_grid()'
to display side-by-side panels to enable easy comparison.
}
\examples{
This will return a paneled plot that shows results for three variance assessments.
\dontrun{p.lmm <- mbecVarianceStatsPlot(variance.obj=list(df1, df2, df3))}
}
\keyword{linear}
\keyword{mixed}
\keyword{models}
\keyword{plot}
\keyword{proportion}
\keyword{variance}
