% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbecs_themes.R
\name{theme_rle}
\alias{theme_rle}
\title{RLE theme for ggplot2 output.}
\usage{
theme_rle(
  base_size = 18,
  base_family = "Myriad Pro",
  legend_position = "bottom"
)
}
\arguments{
\item{base_size}{set base font size - default is 18}

\item{base_family}{set base font family - default is Myriad Pro}

\item{legend_position}{on of "top", "bottom", "left", "right" determines legend position}
}
\value{
A poster theme with legend at the bottom.
}
\description{
Default theme for plotting the packages RLE-plots.
}
\keyword{RLE}
\keyword{theme}
